# docs/swagger.yaml
basePath: /
definitions:
...
  utils.APIResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
  utils.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for MonitoringCR.
  title: MonitoringCR API
  version: "1.0"
paths:
...
  /api/dashboard/due-today:
    get:
      description: Retrieve change requests that are due today, grouped by title,
        PIC, module, and status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get CRs Due Today Statistics
      tags:
      - Charts
  /api/dashboard/kpi-summary:
    get:
      description: Retrieve key performance indicators including total CR, in-progress
        count, near deadline, overdue, cancelled, and completed items.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get KPI Summary Dashboard
      tags:
      - Charts
  /api/dashboard/lifecycle-line-chart:
    get:
      description: Retrieve monthly progression of CRs through ISSUED, IN_PROGRESS,
        and COMPLETE statuses. Supports filtering by year.
      parameters:
      - description: 'Year filter (default: current year)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get CR Lifecycle Line Chart
      tags:
      - Charts
  /api/dashboard/module-category:
    get:
      description: Retrieve cross-tabulation of modules and categories showing the
        count of CRs for each combination.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Module vs Category Statistics
      tags:
      - Charts
...
  /api/dashboard/module-status:
    get:
      description: Retrieve cross-tabulation of modules and statuses showing the count
        of CRs for each combination.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Module vs Status Statistics
      tags:
      - Charts
  /api/dashboard/top-pic:
    get:
      description: Retrieve PIC (Person In Charge) statistics sorted by CR count,
        including in-progress and near-deadline counts with pagination.
      parameters:
      - description: 'Offset (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Limit (default: 5, max: 5)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Top PIC Statistics
      tags:
      - Charts
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"